NNDynamics

Documentation for NNDynamics.

NNDynamics.jl learns a model of the environment. It works with the following environments:

Pendulum

Acrobot

Lunar Lander

Installation

In the julia REPL, run

using Pkg
Pkg.add(url="https://github.com/SvenDuve/NNDynamics.jl")

This package is not for general use. To be fully functional, it requires the following packages:

RLTypes

Conda

PyCall

Gymnasium, this is a slightly adapted version. The original package can be found here

Run the following:

using Pkg
Pkg.add(url="https://github.com/SvenDuve/RLTypes.jl")
Pkg.add("Conda")
Pkg.add("PyCall")

Then again within julia,

using Conda
using PyCall
Conda.pip("install", "git+https://github.com/SvenDuve/Gymnasium")

Bring package into scope with

using NNDynamics

Example usage

Once the above is installed, the following code can be used to train an agent on data collected form 100 episodes of the Pendulum environment:

using NNDynamics
using RLTypes
model(Pendulum(), ModelParameter(collect_train=100, collect_test=10))

For further parameters, see the code.

This function returns a data structure containing a trained model.

NNDynamics.NNModel

NNDynamics.NNModel

NNDynamics.accuracy

NNDynamics.modelEnv

NNDynamics.modelEnv

NNDynamics.train_step!

NNDynamics.train_step!

NNDynamics.NNModel — Method

(m::NNModel)(datapoints)

Functor to apply the NNModel to the given datapoints.

NNDynamics.NNModel — Method

NNModel(input_size::Int, ode_size::Int, output_size::Int)

Create a NNModel with the given input, hidden and output size.

NNDynamics.accuracy — Method

accuracy(y_true, y_pred, tolerance)

Return the accuracy of the prediction.

NNDynamics.modelEnv — Method

modelEnv(environment::ContinuousEnvironment, modelParams::ModelParameter)

Main model training algorithm.

NNDynamics.modelEnv — Method

train_step!(environment::DiscreteEnvironment, S, A, R, S´, T, fθ, model_opt, ep::EnvParameter)

Main algorithm to train the discrete action model.

NNDynamics.train_step! — Method

train_step!(environment::ContinuousEnvironment, S, A, R, S´, T, fθ, model_opt)

Train the continuous action model on gradient descent.

NNDynamics.train_step! — Method

train_step!(environment::DiscreteEnvironment, S, A, R, S´, T, fθ, model_opt, ep::EnvParameter)

Train the discrete action model on gradient descent.